refactor youtube page into components
